+++ src/substruct/substruct_constants.h
@@ -1,0 +1,33 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef NVMOLKIT_SUBSTRUCT_CONSTANTS_H
+#define NVMOLKIT_SUBSTRUCT_CONSTANTS_H
+
+namespace nvMolKit {
+
+//! Maximum size for GPU substruct path
+constexpr int kMaxTargetAtoms = 128;
+constexpr int kMaxQueryAtoms  = 64;
+
+/// Entries per overflow buffer per block
+constexpr int kOverflowEntriesPerBuffer = 2048;
+
+/// Maximum nesting depth for recursive SMARTS patterns
+constexpr int kMaxSmartsNestingDepth = 4;
+
+}  // namespace nvMolKit
+
+#endif  // NVMOLKIT_SUBSTRUCT_CONSTANTS_H
